Move 3D components into 3D private dir
