+++ app/(home)/_3D/CameraAnimation.tsx
