+++ app/(home)/_3D/Rocks3DScene.tsx
