+++ app/(home)/_3D/RocksLighting.tsx
