+++ app/(home)/_3D/RocksStackObject.tsx
